test: add shared request assertion core and reuse in fixtures
Introduce shared request assertion helpers under tests/helpers with:
- request_called_with() for legacy bool checks
- assert_request_called_with() for explicit params/json assertions

Wire tests/conftest.py unifi_called_with fixture to use the shared core
without changing legacy semantics, and convert v1 helper module to a
thin re-export of the shared assertion API.
