+++ tests/conftest.py
@@ -12,6 +12,8 @@ import pytest
 from aiounifi.controller import Controller
 from aiounifi.models.configuration import Configuration
 
+from tests.helpers.request_assertions import request_called_with
+
 
 @pytest.fixture(name="mock_aioresponse")
 def aioresponse_fixture() -> aioresponses:
@@ -44,30 +46,7 @@ def unifi_called_with(mock_aioresponse) -> Callable[[str, str, dict[str, Any]], 
 
     def verify_call(method: str, path: str, **kwargs: dict[str, Any]) -> bool:
         """Verify expected data was provided with a request to aioresponse."""
-        for req, call_list in mock_aioresponse.requests.items():
-            if method != req[0]:
-                continue
-
-            if not req[1].path.endswith(path):
-                continue
-
-            for call in call_list:
-                successful_match = True
-
-                for key, value in kwargs.items():
-                    if key not in call[1] or call[1][key] != value:
-                        successful_match = False
-
-                for key, value in call[1].items():
-                    if key == "allow_redirects":
-                        continue
-                    if value and key not in kwargs:
-                        successful_match = False
-
-                if successful_match:
-                    return True
-
-        return False
+        return request_called_with(mock_aioresponse, method, path, **kwargs)
 
     return verify_call
